+++ CHANGES.rst
@@ -12,6 +12,7 @@ Added
 Fixed
 -----
 - Avoid memory leak from using ``@lru_cache`` on a method.
+- Handle files encoded with an encoding other than UTF-8 without an exception.
 
 
 1.4.2_ - 2022-03-12